requirements.txt
flask
requests
python-dotenv
openai
pydub